+++ bugsnag-android-core/src/test/java/com/bugsnag/android/ConfigurationFacadeTest.java
@@ -91,10 +91,17 @@ public class ConfigurationFacadeTest {
         assertTrue(config.impl.getPersistUser());
     }
 
+    @SuppressWarnings("deprecation")
     @Test
     public void launchCrashThresholdMsValid() {
         config.setLaunchCrashThresholdMs(123456);
-        assertEquals(123456, config.impl.getLaunchCrashThresholdMs());
+        assertEquals(123456, config.impl.getLaunchDurationMillis());
+    }
+
+    @Test
+    public void launchDurationMillisValid() {
+        config.setLaunchDurationMillis(123456);
+        assertEquals(123456, config.impl.getLaunchDurationMillis());
     }
 
     @Test